flarion substr and instr funcs now work
