+++ crates/polars-ops/src/chunked_array/strings/flarion_instr.rs
@@ -1,3 +1,4 @@
+use polars_core::prelude::flarion_funcs::flarion_get_char_position;
 use polars_core::prelude::{Int32Chunked, StringChunked};
 use polars_core::series::{IntoSeries, Series};
 use polars_error::{PolarsError, PolarsResult};
@@ -12,7 +13,7 @@ pub fn flarion_instr_helper(
         let pattern = pattern_series.get(0).unwrap_or("");
         string_series
             .into_iter()
-            .map(|opt_str| opt_str.map(|s| get_char_position(s, pattern)))
+            .map(|opt_str| opt_str.map(|s| flarion_get_char_position(s, pattern)))
             .collect::<Int32Chunked>()
             .into_series()
     } else if string_series.len() == pattern_series.len() {
@@ -20,7 +21,7 @@ pub fn flarion_instr_helper(
             .into_iter()
             .zip(pattern_series)
             .map(|(opt_s, opt_p)| match (opt_s, opt_p) {
-                (Some(s), Some(p)) => Some(get_char_position(s, p)),
+                (Some(s), Some(p)) => Some(flarion_get_char_position(s, p)),
                 _ => None,
             })
             .collect::<Int32Chunked>()
@@ -32,18 +33,4 @@ pub fn flarion_instr_helper(
     };
 
     Ok(result)
-}
-
-fn get_char_position(haystack: &str, needle: &str) -> i32 {
-    if needle.is_empty() {
-        return 1;
-    }
-
-    match haystack.find(needle) {
-        Some(byte_idx) => {
-            // Always add 1 since SQL uses 1-based indexing
-            1 + haystack[..byte_idx].chars().count() as i32
-        },
-        None => 0,
-    }
 }
